// src/_staart/rest/organization.d.ts
export declare const getOrganizationForUser: (userId: string | ApiKeyResponse, organizationId: string) => Promise<import(".prisma/client").organizations>;
export declare const newOrganizationForUser: (userId: string, organization: organizationsCreateInput, locals: Locals) => Promise<import(".prisma/client").organizations>;
export declare const updateOrganizationForUser: (userId: string | ApiKeyResponse, organizationId: string, data: organizationsUpdateInput, locals: Locals) => Promise<import(".prisma/client").organizations>;
export declare const deleteOrganizationForUser: (userId: string | ApiKeyResponse, organizationId: string, locals: Locals) => Promise<void>;
// ...
export declare const getAllOrganizationDataForUser: (userId: string | ApiKeyResponse, organizationId: string) => Promise<{
    billing: import("stripe").Stripe.Customer | import("stripe").Stripe.DeletedCustomer;
    subscriptions: any;
    invoices: any;
    sources: any;
    autoJoinDomain: boolean;
    createdAt: Date;
    forceTwoFactor: boolean;
    id: number;
    ipRestrictions: string | null;
    name: string;
    onlyAllowDomain: boolean;
    profilePicture: string;
    stripeCustomerId: string | null;
    updatedAt: Date;
    username: string;
    assistantName: string;
    assistantSignature: string;
    schedulingDays: string;
    schedulingTimeStart: string;
    schedulingTimeEnd: string;
    schedulingPadding: number;
    schedulingDuration: number;
    schedulingType: "IN_PERSON" | "PHONE_CALL" | "VIDEO_CALL";
    schedulingLocation: number;
    calendars: string | null;
    customEmailEnabled: boolean;
    customEmailAddress: string | null;
    customEmailHost: string | null;
    customEmailPort: number | null;
    customEmailSecure: boolean;
    customEmailUsername: string | null;
    customEmailPassword: string | null;
    memberships: import(".prisma/client").memberships[];
    api_keys: import(".prisma/client").api_keys[];
    domains: import(".prisma/client").domains[];
    webhooks: import(".prisma/client").webhooks[];
} | {
    autoJoinDomain: boolean;
    createdAt: Date;
    forceTwoFactor: boolean;
    id: number;
    ipRestrictions: string | null;
    name: string;
    onlyAllowDomain: boolean;
    profilePicture: string;
    stripeCustomerId: string | null;
    updatedAt: Date;
    username: string;
    assistantName: string;
    assistantSignature: string;
    schedulingDays: string;
    schedulingTimeStart: string;
    schedulingTimeEnd: string;
    schedulingPadding: number;
    schedulingDuration: number;
    schedulingType: "IN_PERSON" | "PHONE_CALL" | "VIDEO_CALL";
    schedulingLocation: number;
    calendars: string | null;
    customEmailEnabled: boolean;
    customEmailAddress: string | null;
    customEmailHost: string | null;
    customEmailPort: number | null;
    customEmailSecure: boolean;
    customEmailUsername: string | null;
    customEmailPassword: string | null;
    memberships: import(".prisma/client").memberships[];
    api_keys: import(".prisma/client").api_keys[];
    domains: import(".prisma/client").domains[];
    webhooks: import(".prisma/client").webhooks[];
}>;